requirements.txt
customtkinter==5.0.3
darkdetect==0.8.0
loguru==0.6.0
numpy==1.24.0
Pillow==9.3.0
tk==0.1.0
